fix: 2FA setup quebrava em PHP 8.1+ (str_pad com float)
Bug pré-existente em lib/totp.php:13 — ceil() retorna float, str_pad()
em PHP 8.1+ exige int. Cast explícito resolve.

Também corrige query do dashboard.php que usava ci.funcionario_id (coluna
não existe — funcionario_id está em assinaturas). Query reescrita com
JOIN assinaturas a ON a.id = ci.assinatura_id.

Remove debug temporários (display_errors em seguranca.php e endpoint
debug_error_log.php) — diagnóstico concluído via log do Hostinger.

+++ lib/totp.php
@@ -10,7 +10,8 @@ function totp_base32_encode(string $bytes): string {
     $alphabet = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ234567';
     $bits = '';
     foreach (str_split($bytes) as $c) $bits .= str_pad(decbin(ord($c)), 8, '0', STR_PAD_LEFT);
-    $bits = str_pad($bits, ceil(strlen($bits) / 5) * 5, '0');
+    // (int) cast: ceil() retorna float, str_pad() em PHP 8.1+ exige int.
+    $bits = str_pad($bits, (int)(ceil(strlen($bits) / 5) * 5), '0');
     $out = '';
     foreach (str_split($bits, 5) as $chunk) $out .= $alphabet[bindec($chunk)];
     return $out;
